requirements.txt
django
djangorestframework
celery
redis-server